Move top-level go files into packager/.
Also, fix a bug in main.go where it wasn't calling StartCron() with the
right number of args.

Fixes #28.
