+++ packager/rtv_test.go
@@ -1,4 +1,4 @@
-package amppackager
+package packager
 
 import (
 	"fmt"
